Extract Header class into separate file
